move test utilities to the cloud provider repository
