+++ pkg/controller/cloud/node_lifecycle_controller_test.go
@@ -32,7 +32,7 @@ import (
 	"k8s.io/client-go/tools/record"
 	fakecloud "k8s.io/cloud-provider/fake"
 	"k8s.io/klog"
-	"k8s.io/kubernetes/pkg/controller/testutil"
+	"k8s.io/client-go/util/testing"
 )
 
 func Test_NodesDeleted(t *testing.T) {
